Update run_docker_interproscan.sh
For compatibility with Micromeda and Genome Properties. Run all analyses. Due to GP v1.0 only using a subset of InterPro we originally only ran analyses for a subset of the databases. This subset approach is what EBI metagenomics did at the time. GP v2.0 covers about all properties in the InterPro so we now run all analyses.